:recycle: [Refactor] main.basic 파일 분리 #main
