+++ src/basic/store/productSelector.js
@@ -1,0 +1,15 @@
+const productSelector = (() => {
+  let selectedId = null;
+
+  return {
+    set: (id) => {
+      selectedId = id;
+    },
+    get: () => selectedId,
+    clear: () => {
+      selectedId = null;
+    },
+  };
+})();
+
+export default productSelector;
